generate HTML manpage for gh-pages, remove options list and cut down history in README
Although our README isn't *that* long as some others projects seem to use it as
the actual documentation, but still contains too much information and, in a
way, duplicated.

By using an HTML version of pipes.sh(6), we can maintain the doc in just one
place, at least for some parts.

----

The HTML generated by `groff` is modified for monospace font (by browser's
setting) and fixed width of 79 characters.