add benchmark script
This script takes/benchmarks revisions of pipes.sh, showing results in
pipe chars per second, an example for the first commit, current tags, and
previous commit (918684b) is:

    $ scripts/benchmark.sh f7d0941 $(git tag) HEAD
    f7d0941   :    564 c/s
    2013-02-01:    534 c/s
    2013-02-06:    534 c/s
    2013-02-07:    540 c/s
    2014-02-20:    531 c/s
    2014-02-21:    552 c/s
    v0.1.0    :    537 c/s
    v0.1.1    :    534 c/s
    v0.2.0    :    526 c/s
    v1.0.0    :    534 c/s
    v1.1.0    :    529 c/s
    v1.2.0    :    529 c/s
    v1.3.0    :    492 c/s
    HEAD      :   3030 c/s

(a dip in v1.3.0 and boost because of not using `tput cup` (c582ec0))

pipes.sh is monkey-patched, so that it doesn't have any delays, that is
replacing `read` with `:`.  Also exiting when reaching `-r LIMIT`, which
is set to 1000.

With execution time (real %e) measured by Bash time builtiin, it
calculates pipe chars per second.